+++ service-skill/src/main/java/com/example/serviceskill/dto/SkillResponse.java
@@ -6,11 +6,11 @@ public record SkillResponse(
         Integer id,
         String name,
         String description,
-        double availableQuantity,
+        Integer availableQuantity,
         BigDecimal price,
-        int nbInscrits,
+        Integer nbInscrits,
         Integer categoryId,
         String categoryName,
         String categoryDescription,
-        Long userId // ID de l'utilisateur (PROVIDER)
+        Long userId
 ) {}
